+++ src/test/java/com/google/firebase/messaging/MessageTest.java
@@ -449,6 +449,72 @@ public class MessageTest {
   }
 
   @Test
+  public void testApnsMessageWithCriticalSound() throws IOException {
+    Message message = Message.builder()
+        .setApnsConfig(ApnsConfig.builder()
+            .setAps(Aps.builder()
+                .setSound(CriticalSound.builder() // All fields
+                    .setCritical(true)
+                    .setName("default")
+                    .setVolume(0.5)
+                    .build())
+                .build())
+            .build())
+        .setTopic("test-topic")
+        .build();
+    Map<String, Object> payload = ImmutableMap.<String, Object>of(
+        "aps", ImmutableMap.builder()
+            .put("sound", ImmutableMap.of(
+                "critical", new BigDecimal(1),
+                "name", "default",
+                "volume", new BigDecimal(0.5)))
+            .build());
+    assertJsonEquals(
+        ImmutableMap.of(
+            "topic", "test-topic",
+            "apns", ImmutableMap.<String, Object>of("payload", payload)),
+        message);
+
+    message = Message.builder()
+        .setApnsConfig(ApnsConfig.builder()
+            .setAps(Aps.builder()
+                .setSound(CriticalSound.builder() // Name field only
+                    .setName("default")
+                    .build())
+                .build())
+            .build())
+        .setTopic("test-topic")
+        .build();
+    payload = ImmutableMap.<String, Object>of(
+        "aps", ImmutableMap.builder()
+            .put("sound", ImmutableMap.of("name", "default"))
+            .build());
+    assertJsonEquals(
+        ImmutableMap.of(
+            "topic", "test-topic",
+            "apns", ImmutableMap.<String, Object>of("payload", payload)),
+        message);
+  }
+
+  @Test
+  public void testInvalidCriticalSound() {
+    List<CriticalSound.Builder> soundBuilders = ImmutableList.of(
+        CriticalSound.builder(),
+        CriticalSound.builder().setCritical(true).setVolume(0.5),
+        CriticalSound.builder().setVolume(-0.1),
+        CriticalSound.builder().setVolume(1.1)
+    );
+    for (int i = 0; i < soundBuilders.size(); i++) {
+      try {
+        soundBuilders.get(i).build();
+        fail("No error thrown for invalid sound: " + i);
+      } catch (IllegalArgumentException expected) {
+        // expected
+      }
+    }
+  }
+
+  @Test
   public void testInvalidApnsConfig() {
     List<ApnsConfig.Builder> configBuilders = ImmutableList.of(
         ApnsConfig.builder(),
@@ -464,22 +530,21 @@ public class MessageTest {
       }
     }
 
-    Aps.Builder builder = Aps.builder().setAlert("string").setAlert(ApsAlert.builder().build());
-    try {
-      builder.build();
-      fail("No error thrown for invalid aps");
-    } catch (IllegalArgumentException expected) {
-      // expected
+    List<Aps.Builder> apsBuilders = ImmutableList.of(
+        Aps.builder().setAlert("string").setAlert(ApsAlert.builder().build()),
+        Aps.builder().setSound("default").setSound(CriticalSound.builder()
+            .setName("default")
+            .build()),
+        Aps.builder().setMutableContent(true).putCustomData("mutable-content", 1)
+    );
+    for (int i = 0; i < apsBuilders.size(); i++) {
+      try {
+        apsBuilders.get(i).build();
+        fail("No error thrown for invalid aps: " + i);
+      } catch (IllegalArgumentException expected) {
+        // expected
+      }
     }
-
-    builder = Aps.builder().setMutableContent(true).putCustomData("mutable-content", 1);
-    try {
-      builder.build();
-      fail("No error thrown for invalid aps");
-    } catch (IllegalArgumentException expected) {
-      // expected
-    }
-
 
     List<ApsAlert.Builder> notificationBuilders = ImmutableList.of(
         ApsAlert.builder().addLocalizationArg("foo"),